feat: Zane's stan example
From our discussion, this is an example of the first type of model (we need to write down our terminology soon and links to all the relevant info). This is a rough stan example that I wrote that probably will not work as-is but demonstrates how I think this type of model should be structured.